Updated random number fragment changed logic to allow for larger (Long) numbers worked on layout

+++ app/src/main/java/com/fimappware/willofthegods/extensions.kt
@@ -21,7 +21,7 @@ fun Context.hideKeyboard(view: View) {
 
 fun String.isNumber(): Boolean {
     return try {
-        this.toInt()
+        this.toLong()
         true
     } catch (ex: NumberFormatException) {
         false
